// src/api/client/masterdata/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type CityViewModel = {
    id?: number;
    name?: (string) | null;
    country?: (string) | null;
};

export type PackageViewModel = {
    packageId?: number;
    packageName?: (string) | null;
};

export type PortViewModel = {
    portId?: number;
    portName?: (string) | null;
    country?: (string) | null;
};

export type ProductViewModel = {
    productId?: number;
    productName?: (string) | null;
};

export type ServiceViewModel = {
    serviceId?: number;
    serviceName?: (string) | null;
    serviceDescription?: (string) | null;
    servicesTypeId?: Array<(number)> | null;
};

export type GetCityResponse = (Array<CityViewModel>);

export type GetCityError = unknown;

export type GetPackageData = {
    query?: {
        containerOnly?: boolean;
    };
};

export type GetPackageResponse = (Array<PackageViewModel>);

export type GetPackageError = unknown;

export type GetPortResponse = (Array<PortViewModel>);

export type GetPortError = unknown;

export type PostPortData = {
    body?: PortViewModel;
};

export type PostPortResponse = (PortViewModel);

export type PostPortError = unknown;

export type GetPortByIdData = {
    path: {
        id: number;
    };
};

export type GetPortByIdResponse = (PortViewModel);

export type GetPortByIdError = (unknown);

export type PutPortByIdData = {
    body?: PortViewModel;
    path: {
        id: number;
    };
};

export type PutPortByIdResponse = (unknown);

export type PutPortByIdError = (unknown);

export type DeletePortByIdData = {
    path: {
        id: number;
    };
};

export type DeletePortByIdResponse = (unknown);

export type DeletePortByIdError = (unknown);

export type GetProductResponse = (Array<ProductViewModel>);

export type GetProductError = unknown;

export type PostProductData = {
    body?: ProductViewModel;
};

export type PostProductResponse = (ProductViewModel);

export type PostProductError = unknown;

export type GetProductByIdData = {
    path: {
        id: number;
    };
};

export type GetProductByIdResponse = (ProductViewModel);

export type GetProductByIdError = (unknown);

export type PutProductByIdData = {
    body?: ProductViewModel;
    path: {
        id: number;
    };
};

export type PutProductByIdResponse = (unknown);

export type PutProductByIdError = (unknown);

export type DeleteProductByIdData = {
    path: {
        id: number;
    };
};

export type DeleteProductByIdResponse = (unknown);

export type DeleteProductByIdError = (unknown);

export type PostServiceData = {
    body?: ServiceViewModel;
};

export type PostServiceResponse = (number);

export type PostServiceError = unknown;

export type GetServiceResponse = (Array<ServiceViewModel>);

export type GetServiceError = unknown;

export type GetServiceByIdData = {
    path: {
        id: number;
    };
};

export type GetServiceByIdResponse = (ServiceViewModel);

export type GetServiceByIdError = (unknown);

export type PutServiceByIdData = {
    body?: ServiceViewModel;
    path: {
        id: number;
    };
};

export type PutServiceByIdResponse = (unknown);

export type PutServiceByIdError = (unknown);

export type DeleteServiceByIdData = {
    path: {
        id: number;
    };
};

export type DeleteServiceByIdResponse = (unknown);

export type DeleteServiceByIdError = (unknown);